Add python3 dependency to ci

+++ hack/release-manifests.sh
@@ -4,7 +4,13 @@ set -e
 PROJECT_ROOT="$(readlink -e $(dirname "$BASH_SOURCE[0]")/../)"
 OUT_DIR=${PROJECT_ROOT}/_out
 
-TAG="${1:-v0.0.0}"
+TAG="${1:-latest}"
+
+if [ "${TAG}" == "latest" ]; then
+  echo "Manifests release will not apply on \"latest\" tag"
+  exit 0
+fi
+
 VERSION=${TAG#v}
 
 rm -rf ${OUT_DIR}
